build: Fix Android and Xcode compilation errors/warnings
Commit f69548c3b3 ("examples: Enable all examples to build on all
platforms") modified dpfp and sam3u_benchmark to be buildable on any
platform, however there were some oversights and regressions introduced
for Android and Xcode.

Update the Android and Xcode build files to account for the removal of
dpfp_threaded.c as well as the inclusion of config.h from examples
and/or tests source.

Additionally switch the threaded version of dpfp to use sem_open()
instead of sem_init() as the latter is in fact deprecated on MacOS.

Closes #808

+++ examples/sam3u_benchmark.c
@@ -64,7 +64,7 @@ static void get_timestamp(struct timeval *tv)
 
 	(void)clock_gettime(CLOCK_MONOTONIC, &ts);
 	tv->tv_sec = ts.tv_sec;
-	tv->tv_usec = ts.tv_nsec / 1000L;
+	tv->tv_usec = (int)(ts.tv_nsec / 1000L);
 #else
 	gettimeofday(tv, NULL);
 #endif
